refactor the intervention mix part
